requirements.txt
flask==3.0.0
flask-cors==4.0.0
ollama==0.1.6
pillow==10.2.0
